chore: refactor the chat-thread-list into it's own component
