+++ packages/chat-component/src/index.ts
@@ -5,5 +5,6 @@ export * from './components/voice-input-button.js';
 export * from './components/document-previewer.js';
 export * from './components/loading-indicator.js';
 export * from './components/citation-list.js';
+export * from './components/chat-thread-component.js';
 export * from './core/index.js';
 export * from './utils/index.js';
